Fixed a race condition during Transport Protocol
A race condition existed between sending data and state managament,
where a recv in response to the send could fly in an update state before
the background task could finish updating state for the message that was
just sent.

This fix re-arranges transmission and state management to ensure the
state is ready for a recv before transmission is sent.